Changed digitMaps data type from byte to int in order to simplify handling more than 127 maps and permit handling more than 255.
git-svn-id: https://svn.code.sf.net/p/controlremote/code/trunk@174 bfc585dc-5255-483b-a5ca-a9bf5897b33d

+++ km/com/hifiremote/jp1/DeviceUpgrade.java
@@ -197,7 +197,7 @@ public class DeviceUpgrade
   {
     System.err.println( "DeviceUpgrade.findDigitIndex()" );
     Button[] buttons = remote.getUpgradeButtons();
-    byte[] digitMaps = remote.getDigitMaps();
+    int[] digitMaps = remote.getDigitMaps();
     if (( digitMaps != null ) && ( protocol.getDefaultCmd().length() == 1 ))
     {
       for ( int i = 0; i < digitMaps.length; i++ )
